% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn_fauna.R
\name{bn_species_fit}
\alias{bn_species_fit}
\title{called by fit_bn_fauna, Fits BN to subset of data for a given species
bn_str: BN structure can be fixed}
\usage{
bn_species_fit(bn_data, bn_str)
}
\description{
called by fit_bn_fauna, Fits BN to subset of data for a given species
bn_str: BN structure can be fixed
}
