% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn.R
\name{fit_bn}
\alias{fit_bn}
\title{Fits a BN using covariates from rasters}
\usage{
fit_bn(geom, vars_paths, crs = "lcc_mex", score = "bic")
}
\arguments{
\item{geom}{region sf, polygon(s) that determine analysis area}

\item{vars_paths}{paths to rasters with covariates to be included in BN}

\item{crs}{string specifying projection (proj4string) passed to
sf::st_transform, defaults to lcc (Mex)}

\item{score}{a character string, the label of the network score to be used in
hill-climbing algorithm, pased to \code{\link[bnlearn]{hc}}. Defaults to BIC}
}
\description{
Fits a BN using covariates from rasters
}
\details{
Given the path to a set of rasters and a region
specified by a polygon or set of polygons (simple feature) it adjusts a
Bayesian Network.
}
