% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn_fauna.R
\name{fit_bn_fauna}
\alias{fit_bn_fauna}
\title{Fits a BN for each species on a shapefile using covariates from rasters}
\usage{
fit_bn_fauna(fauna_geom, species_var, geom, covs_paths, crs = "lcc_mex")
}
\arguments{
\item{fauna_geom}{fauna sf, point geometry indicating where species where
observed}

\item{species_var}{unquoted variable indicating column on fauna data that
indicates species}

\item{geom}{region sf, polygon(s) that determine analysis area}

\item{covs_paths}{paths to rasters with covariates to be included in BN}
}
\description{
Fits a BN for each species on a shapefile using covariates from rasters
}
\details{
Given a set of points observations of fauna species and a region
specified by a polygon or set of polygons (simple feature) it adjusts a
Bayesian Network per species and covariates given in rasters.
}
