% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn_fauna.R
\name{prep_geom_covs}
\alias{prep_geom_covs}
\title{Creates a data.frame with covariates from rasters.}
\usage{
prep_geom_covs(covs_paths, geom)
}
\description{
Reads tif files of discretized covariates and id variable,
cuts rasters with polygons in geom, returns data.frame with numerical
variables (except for coordinates and id)
covs_paths paths to rasters with covariates to consider
geom sf to cut and mask brick of covariates
}
