% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn_fauna.R
\name{prep_geom_data}
\alias{prep_geom_data}
\title{called by fit_bn_fauna, prepares data to fit BNs or to evaluate
likelihood
returns a list with an entry per species, each entry contains a data
frame of categorized covarites suitable for fitting/evaluating a BN}
\usage{
prep_geom_data(fauna_geom, species_var, geom, covs_paths, crs)
}
\description{
called by fit_bn_fauna, prepares data to fit BNs or to evaluate
likelihood
returns a list with an entry per species, each entry contains a data
frame of categorized covarites suitable for fitting/evaluating a BN
}
