% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geom_vars.R
\name{prep_geom_vars}
\alias{prep_geom_vars}
\title{Creates a data.frame with covariates from rasters.}
\usage{
prep_geom_vars(vars_paths, geom, crs = "lcc_mex")
}
\arguments{
\item{vars_paths}{paths to rasters with covariates to consider, must include
raster with id's (name of layer should contain "id")}

\item{geom}{region sf, polygon(s) that determine analysis area}

\item{crs}{string specifying projection (proj4string) passed to
sf::st_transform, defaults to lcc (Mex)}
}
\description{
Creates a data.frame with covariates from rasters.
}
\details{
Reads tif files of discretized variables and id variable,
cuts rasters with polygons in a polygon simple feature and returns a
data.frame with the categorical variables, coordinates and id variable.
}
