% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shps.R
\name{read_shps}
\alias{read_shps}
\title{Reads shapefiles, projects to given crs and returns polygon simple feature}
\usage{
read_shps(shp_paths, crs = "lcc_mex", top = TRUE)
}
\arguments{
\item{shp_paths}{vector of path to shape files (or single path)}

\item{crs}{string specifying projection (proj4string) passed to
sf::st_transform, defaults to lcc (Mex)}

\item{top}{in case of multipolygon geometry in one of the single shapefiles,
TRUE indicates that only the biggest polygon will be included, if FALSE all
polygons will be included}
}
\description{
Reads shapefiles, projects to given crs and returns polygon simple feature
}
\details{
Recieves path(s) to shapefile(s), projects to given crs and returns
a single simple feature.
}
