% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shp_point.R
\name{read_shps_points}
\alias{read_shps_points}
\title{Reads shapefiles, projects to given crs and returns point simple feature}
\usage{
read_shps_points(shp_paths, crs = "lcc_mex", filter_var, filter_values)
}
\arguments{
\item{shp_paths}{vector of path to shape files (or single path)}

\item{crs}{string specifying projection (proj4string) passed to
sf::st_transform, defaults to lcc (Mex)}

\item{filter_var}{optional, if filtering by a variable (for example year)
unquoted variable corresponding to the variable name}

\item{filter_values}{optional, values to keep of filter_var}
}
\description{
Recieves path(s) to shapefile(s), projects to given crs and returns a single
simple feature.
}
