% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bn_fauna.R
\name{test_bn_fauna}
\alias{test_bn_fauna}
\title{Tests a BN for each species on a shapefile using covariates from rasters}
\usage{
test_bn_fauna(bn_fitted_list, fauna_geom, species_var, geom, covs_paths,
  crs = "lcc_mex")
}
\arguments{
\item{fauna_geom}{fauna sf, point geometry indicating where species where
observed}

\item{species_var}{unquoted variable indicating column on fauna data that
indicates species}

\item{geom}{region sf, polygon(s) that determine test area}

\item{covs_paths}{paths to rasters with covariates to be included in BN}
}
\description{
Tests a BN for each species on a shapefile using covariates from rasters
}
\details{
Given a set of points observations of fauna species and a a region
specified by a polygon or set of polygons (simple feature) and a fiited
Bayesian Network per species it evaluates the likelihood given the covariates
given in rasters, for the fauna and the no fauna data.
}
