% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_discretize.R
\name{tif_discretize}
\alias{tif_discretize}
\title{Read tif file and discretize values}
\usage{
tif_discretize(tif_path, breaks = 4, extent, path_save, vals_na = -1)
}
\arguments{
\item{tif_path}{path to tif file that will be read and discretized}

\item{breaks}{number of breaks passed to arules::discretize}

\item{extent}{extent to which raster will be cropped and set}

\item{path_save}{path to new file, where the discretized raster will be}

\item{vals_na}{values to be set to NA
saved}
}
\description{
Recieves a path to a tif, reads the raster, crops and sets to given extent,
discretizes values and writes a new raster with the same name in given path
}
