% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bn_str.R
\name{fit_bn_str}
\alias{fit_bn_str}
\title{If structure not given, learn a network structure and fits BN}
\usage{
fit_bn_str(bn_data, bn_str, score = "k2")
}
\arguments{
\item{bn_data}{data.frame with only categorical columns}

\item{bn_str}{optional bayesian network structure, if missing will be learned}

\item{score}{used for structure learning algorthm (hill-climbing)}
}
\description{
If structure not given, learn a network structure and fits BN
}
